+++ contracts/MilestonePricing.sol
@@ -121,7 +121,7 @@ contract MilestonePricing is PricingStrategy, Ownable, SafeMathLib {
   }
 
   /// @dev Calculate the current price for buy in amount.
-  function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint decimals) public constant returns (uint) {
+  function calculatePrice(uint value, uint weiRaised, uint tokensSold, address msgSender, uint8 decimals) public constant returns (uint) {
 
     uint multiplier = 10 ** decimals;
